[hive] Make paimon-hive connecotr no longer dependent on flink
